+++ lib/stormpath-sdk/resource/group_membership.rb
@@ -30,7 +30,21 @@ module Stormpath
         get_resource_property GROUP, Group
       end
 
-      def self.create account, group, data_store
+      def delete
+        data_store.delete self
+      end
+
+      #
+      # THIS IS NOT PART OF THE STORMPATH PUBLIC API.  SDK end-users should not call it - it could be removed or
+      # changed at any time.  It is publicly accessible only as an implementation technique to be used by other
+      # resource classes.
+      #
+      # @param account the account to associate with the group.
+      # @param group the group which will contain the account.
+      # @param data_store the datastore used to create the membership
+      # @return the created GroupMembership instance.
+      #
+      def self._create account, group, data_store
 
         #TODO: enable auto discovery
         href = "/groupMemberships"
@@ -49,10 +63,6 @@ module Stormpath
 
         data_store.create href, group_membership, GroupMembership
 
-      end
-
-      def delete
-        data_store.delete self
       end
 
     end