+++ CHANGES.md
@@ -7,7 +7,7 @@ Version 0.4.0
 Released on October XY, 2012
 
 - The Stormpath::Resource::GroupMembership class now extends the Stormpath::Resource::Resource class. It is no longer possible to call 'save' on an instance of this class.
-- The 'create' method of the Stormpath::Resource::GroupMembership class is now a class method and receives an instance of Stormpath::DataStore::DataStore.
+- The 'create' method of the Stormpath::Resource::GroupMembership class is now a class method and receives an instance of Stormpath::DataStore::DataStore; it was renamed from 'create' to '_create'.
 - The 'add_group' method implementation of the Stormpath::Resource::Account class was updated to reflect the previously mentioned changes.
 - The 'add_account' method implementation of the Stormpath::Resource::Group class was updated to reflect the previously mentioned changes.
 - The 'set_account' and 'set_group' methods were removed from the Stormpath::Resource::GroupMembership class.